Combine multiple IPUMS extract definitions into one

Create a single extract definition that includes all of the specifications included in a set of ipums_extract objects of the same collection.

Learn more about the IPUMS API in vignette("ipums-api").

combine_extracts(...)

Arguments

...

Arbitrary number of ipums_extract objects to be combined. All extracts must belong to the same collection.

Value

An ipums_extract object of the same collection as the extracts provided in ... containing the combined extract definition.

Details

Values that exist in more than one of the provided extract definitions will be de-duplicated such that only one entry for the duplicated specification is included in the final extract.

Some extract fields can only take a single value. In the event that the definitions being combined each have different values for such a field, the value found in the first extract definition provided to ... is retained in the final extract.

To modify values in a single extract definition, see add_to_extract() or remove_from_extract().

See also

Examples

ext1 <- define_extract_nhgis(
  datasets = "2011_2015_ACS5a",
  data_tables = "B00001",
  geog_levels = "county"
)

ext2 <- define_extract_nhgis(
  datasets = c("2011_2015_ACS5a", "2012_2016_ACS5a"),
  data_tables = c("B00002", "B01001"),
  geog_levels = "county"
)

combine_extracts(ext1, ext2)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 2011_2015_ACS5a
#>   Tables: B00001, B00002, B01001
#>   Geog Levels: county
#> 
#> Dataset: 2012_2016_ACS5a
#>   Tables: B00002, B01001
#>   Geog Levels: county