Define an IPUMS extract object

Specify the parameters for a new IPUMS extract request object to be submitted via the IPUMS API. An extract request contains the specifications required to obtain a particular set of data from an IPUMS collection.

Learn more about the IPUMS API in vignette("ipums-api").

Supported collections

Currently, ipumsr supports extract definitions for the following collections:

IPUMS USA: define_extract_usa()

IPUMS CPS: define_extract_cps()

IPUMS NHGIS: define_extract_nhgis()

Value

These functions produce an ipums_extract object with a subclass based on the collection corresponding to the extract request. For instance, an IPUMS NHGIS extract created with define_extract_nhgis() will return an object of classes nhgis_extract and ipums_extract. These objects are compatible with the rest of the API functionality provided by ipumsr.