Define an IPUMS CPS extract request

Define an IPUMS CPS extract request to be submitted via the IPUMS API. An extract request contains the specifications required to identify and obtain a particular set of data from the IPUMS CPS system.

Learn more about the IPUMS API in vignette("ipums-api").

define_extract_cps(
  description,
  samples,
  variables,
  data_format = c("fixed_width", "csv", "stata", "spss", "sas9"),
  data_structure = "rectangular",
  rectangular_on = "P"
)

Arguments

description

Description of the extract.

samples

Character vector of samples to include in the extract. Samples should be specified using IPUMS CPS sample ID values.

variables

Character vector of variables to include in the extract.

data_format

The desired format of the extract data file. One of "fixed_width", "csv", "stata", "spss", or "sas9".

data_structure

The desired structure of the extract data.

"rectangular" provides person records with all requested household information attached to respective household members.

"hierarchical" provides household records followed by person records.

Currently, only "rectangular" extracts are supported.

rectangular_on

Records on which to rectangularize. "P" rectangularizes on person records. "H" provides household-only records.

Currently, only "P" is supported.

Value

An object of class cps_extract containing the extract definition.

See also

Examples

my_extract <- define_extract_cps(
  description = "Example CPS extract definition",
  samples = c("cps2020_02s", "cps2020_03s"),
  variables = c("AGE", "SEX", "YEAR")
)

my_extract
#> Unsubmitted IPUMS CPS extract 
#> Description: Example CPS extract definition
#> Samples: (2 total) cps2020_02s, cps2020_03s
#> Variables: (3 total) AGE, SEX, YEAR

if (FALSE) {
# Use the extract definition to submit an extract request to the API
submit_extract(my_extract)
}