Define an IPUMS NHGIS extract request

Define an IPUMS NHGIS request to be submitted via the IPUMS API. An extract request contains the specifications required to obtain a particular set of data from the IPUMS NHGIS system.

Use get_nhgis_metadata() to browse and identify data sources for use in NHGIS extract definitions.

See Details section for information about creating valid extract definitions. For general information, see the NHGIS data source overview and the FAQ.

Learn more about the IPUMS API in vignette("ipums-api").

define_extract_nhgis(
  description = "",
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  breakdown_and_data_type_layout = NULL,
  tst_layout = NULL,
  data_format = NULL
)

Arguments

description

Description of the extract.

datasets

Datasets to include in the extract request.

data_tables

Summary tables to retrieve for each of the requested datasets. Required for all datasets in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets.

time_series_tables

Time series tables to include in the extract request.

geog_levels

Geographic levels (e.g. "county" or "state") at which to obtain data for the requested datasets or time_series_tables. Required for all datasets and time_series_tables in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets and/or time_series_tables.

years

Years for which to obtain the data contained in the requested datasets. Use "*" to select all available years for a given dataset. Use get_nhgis_metadata() to determine if a dataset allows year selection.

See Details section for syntax options when an extract definition has multiple datasets.

breakdown_values

Breakdown values to apply to the requested datasets.

See Details section for syntax options when an extract definition has multiple datasets.

geographic_extents

Vector of geographic extents to use for all of the datasets in the extract definition (for instance, to obtain data within a particular state). Use "*" to select all available extents.

Required when the extract definition includes any geog_levels that require extent selection. See get_nhgis_metadata() to determine if a geographic level requires extent selection. At the time of writing, NHGIS supports extent selection only for blocks and block groups.

shapefiles

Shapefiles to include in the extract request.

breakdown_and_data_type_layout

The desired layout of any datasets that have multiple data types or breakdown values.

"single_file" (default) keeps all data types and breakdown values in one file

"separate_files" splits each data type or breakdown value into its own file

Required if any datasets included in the extract definition consist of multiple data types (for instance, estimates and margins of error) or have multiple breakdown values specified. See get_nhgis_metadata() to determine whether a requested dataset has multiple data types.

tst_layout

The desired layout of all time_series_tables included in the extract definition. One of:

"time_by_column_layout" (wide format, default): rows correspond to geographic units, columns correspond to different times in the time series

"time_by_row_layout" (long format): rows correspond to a single geographic unit at a single point in time

"time_by_file_layout": data for different times are provided in separate files

Required when an extract definition includes any time_series_tables.

data_format

The desired format of the extract data file. One of:

"csv_no_header" (default) includes only a minimal header in the first row

"csv_header" includes a second, more descriptive header row. (By default this header is removed when reading data with read_nhgis(), but it can be retained.)

"fixed_width" provides data in a fixed width format

Required when an extract definition includes any datasets or time_series_tables.

Value

An object of class nhgis_extract containing the extract definition.

Details

An NHGIS extract definition must include at least one dataset, time series table, or shapefile. For extract definitions that include any datasets, corresponding data_tables and geog_levels must be specified. For extract definitions that include any time_series_tables, corresponding geog_levels must be specified.

NHGIS extract definitions may contain multiple datasets or time_series_tables. Each dataset or time series table is associated with several subfields that apply only to that particular dataset or time series table.

The following arguments are applied to the requested datasets:

data_tables

years

breakdown_values

The following arguments are applied to the requested datasets and time_series_tables:

geog_levels

To apply the values of these arguments to all of the datasets and/or time_series_tables in the extract definition, pass the values as a vector:

define_extract_nhgis(
  datasets = c("2016_2020_ACS5a", "2017_2021_ACS5a"),
  data_tables = "B01001",
  geog_levels = c("county", "state")
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 2016_2020_ACS5a
#>   Tables: B01001
#>   Geog Levels: county, state
#> 
#> Dataset: 2017_2021_ACS5a
#>   Tables: B01001
#>   Geog Levels: county, state

To apply the values of these arguments to specific datasets and/or time_series_tables in the extract definition, pass the values as a named list. In this case, names should correspond to the datasets and/or time_series_tables and their associated values will only be applied to the indicated dataset or time series table:

define_extract_nhgis(
  datasets = "2016_2020_ACS5a",
  data_tables = "B01001",
  time_series_tables = "A00",
  geog_levels = list(
    "2016_2020_ACS5a" = "county",
    "A00" = "state"
  )
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 2016_2020_ACS5a
#>   Tables: B01001
#>   Geog Levels: county
#> 
#> Time Series Table: A00
#>   Geog Levels: state

Note that while it is possible to match elements to datasets and/or time_series_tables by index by providing an unnamed list, we recommend using names to explicitly specify the values that should be applied to each of the request's datasets and time_series_tables.

Alternatively, you can create multiple extract definitions and combine them with combine_extracts().

See also

Examples

# Extract definition for tables from an NHGIS dataset
my_extract <- define_extract_nhgis(
  description = "Example NHGIS extract",
  datasets = "1990_STF3",
  data_tables = "NP57",
  geog_levels = c("county", "tract")
)

my_extract
#> Unsubmitted IPUMS NHGIS extract 
#> Description: Example NHGIS extract
#> 
#> Dataset: 1990_STF3
#>   Tables: NP57
#>   Geog Levels: county, tract

# Extract definition for an NHGIS time series table
define_extract_nhgis(
  description = "Example NHGIS extract",
  time_series_tables = "CL8",
  geog_levels = "county"
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: Example NHGIS extract
#> 
#> Time Series Table: CL8
#>   Geog Levels: county

# Use get_nhgis_metadata() to identify data source names
if (FALSE) {
get_nhgis_metadata("time_series_tables")
}

# Fields that are attached to specific datasets/time series tables
# are recycled to all datasets/time series tables in an extract by default
define_extract_nhgis(
  description = "Extract definition with multiple datasets",
  datasets = c("2014_2018_ACS5a", "2015_2019_ACS5a"),
  data_tables = "B01001",
  geog_levels = c("state", "county")
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: Extract definition with multiple datasets
#> 
#> Dataset: 2014_2018_ACS5a
#>   Tables: B01001
#>   Geog Levels: state, county
#> 
#> Dataset: 2015_2019_ACS5a
#>   Tables: B01001
#>   Geog Levels: state, county

# To instead attach specific values to each dataset or time series table,
# use a named list instead of a vector:
define_extract_nhgis(
  description = "Extract with multiple time series tables",
  time_series_tables = c("CW3", "CW5"),
  geog_levels = list(CW5 = "county", CW3 = "state")
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: Extract with multiple time series tables
#> 
#> Time Series Table: CW3
#>   Geog Levels: state
#> 
#> Time Series Table: CW5
#>   Geog Levels: county

# For extracts with datasets and time series tables,
# geog_levels are matched to both sources:
define_extract_nhgis(
  description = "Extract with datasets and time series tables",
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  time_series_tables = "CL6",
  geog_levels = "state"
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: Extract with datasets and time series tables
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2
#>   Geog Levels: state
#> 
#> Time Series Table: CL6
#>   Geog Levels: state

define_extract_nhgis(
  description = "Extract with datasets and time series tables",
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  time_series_tables = "CL6",
  geog_levels = list("1990_STF1" = "county", "CL6" = "state")
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: Extract with datasets and time series tables
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2
#>   Geog Levels: county
#> 
#> Time Series Table: CL6
#>   Geog Levels: state

if (FALSE) {
# Use the extract definition to submit an extract request to the API
submit_extract(my_extract)
}