Convert recent extract definitions from tibble to list format

Convert a tibble of extract definition specifications to a list of ipums_extract objects or vice versa.

Use get_extract_info() to obtain information on recent extracts.

Learn more about the IPUMS API in vignette("ipums-api").

extract_tbl_to_list(extract_tbl, validate = TRUE)

extract_list_to_tbl(extract_list)

Arguments

extract_tbl

A tibble (or data.frame) containing the specifications for one or more ipums_extract objects.

Use get_extract_info() with table = TRUE to produce such an object.

validate

Logical value indicating whether to check that each of the output ipums_extract objects contains a valid and complete extract definition. Defaults to TRUE.

extract_list

A list of ipums_extract objects or a single ipums_extract object.

Value

For extract_tbl_to_list(), a list of length equal to the number of extracts represented in extract_tbl.

For extract_list_to_tbl(), a tibble

representing the specifications for each of the extract requests represented in extract_list. Each column corresponds to an extract field.

Examples

if (FALSE) {
# Get tibble of recent extracts
tbl_of_last_10_extracts <- get_extract_info("usa", table = TRUE)

# Filter down to extracts with "income" in the description
description_mentions_income <- grepl(
  "[Ii]ncome",
  tbl_of_last_10_extracts$description
)

income_extracts <- tbl_of_last_10_extracts[description_mentions_income, ]

# Convert tibble of extracts to list of extracts
income_extracts <- extract_tbl_to_list(income_extracts)

# Now it's easier to operate on those elements as extract objects:
revised_income_extract <- add_to_extract(
  income_extracts[[1]],
  samples = "us2018a"
)

submitted_revised_income_extract <- submit_extract(revised_income_extract)

# Get list of recent extracts
list_of_last_10_extracts <- get_extract_info("usa")

# Print the extract number for extracts that are downloadable:
for (extract in list_of_last_10_extracts) {
  if (is_extract_ready(extract)) print(extract$number)
}

# Convert list of extracts to tibble of extracts to view in a tabular format
extract_list_to_tbl(list_of_last_10_extracts)
}