Get IPUMS file information

Get IPUMS metadata information about the data file loaded into R from an ipums_ddi

ipums_file_info(object, type = NULL)

Arguments

object
An ipums_ddi object (loaded with read_ipums_ddi).

type
NULL to load all types, or one of "ipums_project", "extract_data", "extract_notes", "conditions" or "citation".

Value

If type is NULL, a list with the ipums_project, extract_date, extract_notes, conditions, and citation. Otherwise a string with the type of information requested in type.

Examples

ddi <- read_ipums_ddi(ipums_example("cps_00006.xml"))
ipums_file_info(ddi)

#> $ipums_project
#> [1] "IPUMS-CPS"
#> 
#> $extract_date
#> [1] "2017-08-02"
#> 
#> $extract_notes
#> [1] "User-provided description:  Minimal test extract\nSamples: 1962, 1963\nVariables: STATEFIP, INCTOT (automatically Year, SERIAL, HWTSUPP, MONTH, WTSUPP)\nSelect Cases: State - Minnesota, Iowa, Wisconsin, South Dakota, North Dakota"
#> 
#> $conditions
#> [1] "Users of IPUMS-CPS data must agree to abide by the conditions of use. A user's license is valid for one year and may be renewed.  Users must agree to the following conditions:\n\n(1) No fees may be charged for use or distribution of the data.  All persons are granted a limited license to use these data, but you may not charge a fee for the data if you distribute it to others.\n\n(2) Cite IPUMS appropriately.  For information on proper citation,  refer to the citation requirement section of this DDI document.\n\n(3) Tell us about any work you do using the IPUMS.  Publications, research  reports, or presentations making use of IPUMS-CPS should be added to our  Bibliography. Continued funding for the IPUMS depends on our ability to  show our sponsor agencies that researchers are using the data for productive  purposes.\n\n(4) Use it for GOOD -- never for EVIL."
#> 
#> $citation
#> [1] "Publications and research reports based on the IPUMS-CPS database must cite it appropriately. The citation should include the following:\n\nSarah Flood, Miriam King, Steven Ruggles, and J. Robert Warren. Integrated Public Use Microdata Series, Current Population Survey: Version 5.0 [dataset]. Minneapolis, MN: University of Minnesota, 2017. https://doi.org/10.18128/D030.V5.0\n\nThe licensing agreement for use of IPUMS-CPS data requires that users supply us with the title and full citation for any publications, research reports, or educational materials making use of the data or documentation. Please add your citation to the IPUMS bibliography: http://bibliography.ipums.org/"
#>