ipumsr: An R Interface for Downloading, Reading, and Handling IPUMS Data

An easy way to work with census, survey, and geographic data provided by IPUMS in R. Generate and download data through the IPUMS API and load IPUMS files into R with their associated metadata to make analysis easier. IPUMS data describing 1.4 billion individuals drawn from over 750 censuses and surveys is available free of charge from the IPUMS website https://www.ipums.org.

See also

Author

Maintainer: Derek Burk ipums+cran@umn.edu

Authors:

Greg Freedman Ellis

Finn Roberts

Other contributors:

Joe Grover [contributor]

Dan Ehrlich [contributor]

Renae Rodgers [contributor]

Institute for Social Research and Data Innovation ipums@umn.edu [copyright holder]