Create an ipums_ddi object

Constructor to make a new ipums_ddi object.

This function has been deprecated because it is a low-level function that is unlikely to be useful externally. Use read_ipums_ddi() (for microdata projects) or read_nhgis_codebook() (for NHGIS) to create an ipums_ddi object.

make_ddi_from_scratch(
  file_name = NULL,
  file_path = NULL,
  file_type = NULL,
  ipums_project = NULL,
  extract_date = NULL,
  extract_notes = NULL,
  rectypes = NULL,
  rectype_idvar = NULL,
  rectypes_keyvars = NULL,
  var_info = NULL,
  conditions = NULL,
  citation = NULL,
  file_encoding = NULL
)