Remove values from an existing IPUMS USA extract definition

Remove existing values from an IPUMS USA extract definition. All fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS USA extract definition, see add_to_extract(). When replacing values, it is best to first add new values using add_to_extract() before removing values with remove_from_extract(). This limits the possibility of producing a temporarily invalid extract specification.

Learn more about the IPUMS API in vignette("ipums-api").

# S3 method for usa_extract
remove_from_extract(extract, samples = NULL, variables = NULL, ...)

Arguments

extract

An ipums_extract object.

samples

Character vector of samples to remove from the extract.

variables

Character vector of variables to remove from the extract.

...

Ignored

Value

A modified usa_extract object

Details

If the supplied extract definition comes from a previously submitted extract request, this function will reset the definition to an unsubmitted state.

See also

Examples

usa_extract <- define_extract_usa(
  description = "USA example",
  samples = c("us2013a", "us2014a"),
  variables = c("AGE", "SEX", "YEAR")
)

remove_from_extract(
  usa_extract,
  samples = "us2014a",
  variables = c("AGE", "SEX")
)
#> Unsubmitted IPUMS USA extract 
#> Description: USA example
#> Samples: (1 total) us2013a
#> Variables: (1 total) YEAR

# To replace values, use add_to_extract() first to avoid invalid
# extract definitions:
revised_extract <- add_to_extract(usa_extract, samples = "us2015a")
remove_from_extract(revised_extract, samples = c("us2013a", "us2014a"))
#> Unsubmitted IPUMS USA extract 
#> Description: USA example
#> Samples: (1 total) us2015a
#> Variables: (3 total) AGE, SEX, YEAR