Add values to an existing IPUMS NHGIS extract definition

Add new values to an IPUMS NHGIS extract definition. All fields are optional, and if omitted, will be unchanged. Supplying a value for fields that take a single value, such as description and data_format, will replace the existing value with the supplied value.

In general, adding to an extract follows the same syntax conventions as used in define_extract_nhgis(). See Details section for more information on how values passed to dataset and time series table subfields are interpreted.

To remove existing values from an IPUMS NHGIS extract definition, use remove_from_extract(). To add values contained in another extract definition, use combine_extracts().

Learn more about the IPUMS API in vignette("ipums-api").

# S3 method for nhgis_extract
add_to_extract(
  extract,
  description = NULL,
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  breakdown_and_data_type_layout = NULL,
  tst_layout = NULL,
  data_format = NULL,
  ...
)

Arguments

extract

An ipums_extract object.

description

Description of the extract.

datasets

Datasets to include in the extract request.

data_tables

Summary tables to retrieve for each of the requested datasets. If no datasets are provided, applies to those that already exist in the extract definition. Required if any new datasets are specified.

See Details section for syntax options when an extract definition has multiple datasets.

time_series_tables

Time series tables to include in the extract request.

geog_levels

Geographic levels (for example, "county" or "state") at which to obtain data for the requested datasets and time_series_tables. If no datasets and/or time_series_tables are provided, applies to those that already exist in the extract definition. Required if any new datasets or time_series_tables are specified.

See Details section for syntax options when an extract definition has multiple datasets or time_series_tables.

years

Years for which to obtain the data contained in the requested datasets. If no datasets are provided, applies to those that already exist in the extract definition.

Use "*" to select all available years for a given dataset. Use get_nhgis_metadata() to determine if a dataset allows year selection.

See Details section for syntax options when an extract definition has multiple datasets.

breakdown_values

Breakdown values to to apply to the requested datasets. If no datasets are provided, applies to those that already exist in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets.

geographic_extents

Vector of geographic extents to use for all of the datasets in the extract definition (for instance, to obtain data within a particular state). Use "*" to select all available extents.

Required when the extract definition includes any geog_levels that require extent selection. See get_nhgis_metadata() to determine if a geographic level requires extent selection. At the time of writing, NHGIS supports extent selection only for blocks and block groups.

shapefiles

Shapefiles to include in the extract request.

breakdown_and_data_type_layout

The desired layout of any datasets that have multiple data types or breakdown values.

"single_file" (default) keeps all data types and breakdown values in one file

"separate_files" splits each data type or breakdown value into its own file

Required if any datasets included in the extract definition consist of multiple data types (for instance, estimates and margins of error) or have multiple breakdown values specified. See get_nhgis_metadata() to determine whether a requested dataset has multiple data types.

tst_layout

The desired layout of all time_series_tables included in the extract definition. One of:

"time_by_column_layout" (wide format, default): rows correspond to geographic units, columns correspond to different times in the time series

"time_by_row_layout" (long format): rows correspond to a single geographic unit at a single point in time

"time_by_file_layout": data for different times are provided in separate files

Required when an extract definition includes any time_series_tables.

data_format

The desired format of the extract data file. One of:

"csv_no_header" (default) includes only a minimal header in the first row

"csv_header" includes a second, more descriptive header row. (By default this header is removed when reading data with read_nhgis(), but it can be retained.)

"fixed_width" provides data in a fixed width format

Required when an extract definition includes any datasets or time_series_tables.

...

Ignored

Value

A modified nhgis_extract object

Details

For extract definitions that include any datasets, corresponding data_tables and geog_levels must be specified. For extract definitions that include any time_series_tables, corresponding geog_levels must be specified.

NHGIS extract definitions may contain multiple datasets or time_series_tables. Each dataset or time series table is associated with several subfields that apply only to that particular dataset or time series table.

The following arguments are applied to an extract definition's datasets:

data_tables

years

breakdown_values

The following arguments are applied to an extract definition's datasets and time_series_tables:

geog_levels

To add the values of these arguments to multiple datasets and/or time_series_tables in the extract definition, pass the values as a vector. In this case, if any values are provided in the datasets or time_series_tables arguments, subfield values will only be applied to those datasets and/or time_series_tables. Otherwise, they will be applied to all datasets and/or time_series_tables that exist in the extract definition.

To add the values of these arguments to specific datasets and/or time_series_tables in the extract definition, pass the values as a named list. In this case, names should correspond to the datasets and/or time_series_tables and their associated values will only be added to the indicated dataset or time series table. Values can be added for new datasets and time_series_tables or those that already exist in the extract definition.

See examples for demonstrations of this syntax.

Note that while it is possible to match elements to datasets and/or time_series_tables by index by providing an unnamed list, we recommend using names to explicitly specify the values that should be added to each of the request's datasets and time_series_tables.

For extract fields that take a single value, add_to_extract() will replace the existing value with the new value provided for that field. It is not necessary to first remove this value using remove_from_extract().

If the supplied extract definition comes from a previously submitted extract request, this function will reset the definition to an unsubmitted state.

See also

Examples

extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  geog_levels = "county"
)

# Modify existing datasets in extract:
extract2 <- add_to_extract(
  extract,
  data_tables = "NP3"
)

extract2
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2, NP3
#>   Geog Levels: county

# Add a new dataset or time series table. Fields that apply to each dataset
# or time series table are only applied to the newly added dataset or
# time series table:
extract <- add_to_extract(
  extract,
  datasets = "1980_STF1",
  data_tables = "NT1A",
  geog_levels = c("county", "state")
)

add_to_extract(
  extract,
  time_series_tables = "A00",
  geog_levels = "state"
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2
#>   Geog Levels: county
#> 
#> Dataset: 1980_STF1
#>   Tables: NT1A
#>   Geog Levels: county, state
#> 
#> Time Series Table: A00
#>   Geog Levels: state

# To modify existing datasets, use a named list to specify which datasets
# should be attached to which `data_tables` or `geog_levels`.
# Vector arguments are attached to all datasets.
add_to_extract(
  extract,
  data_tables = list(
    "1990_STF1" = "NP4",
    "1980_STF1" = "NT1B"
  ),
  geog_levels = "nation"
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2, NP4
#>   Geog Levels: county, nation
#> 
#> Dataset: 1980_STF1
#>   Tables: NT1A, NT1B
#>   Geog Levels: county, state, nation

# Values that can only take a single value are replaced
add_to_extract(extract, data_format = "fixed_width")$data_format
#> [1] "fixed_width"

# You can also combine two extract requests together
combine_extracts(extract, extract2)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2, NP3
#>   Geog Levels: county
#> 
#> Dataset: 1980_STF1
#>   Tables: NT1A
#>   Geog Levels: county, state