ipums_extract class

The ipums_extract class provides a data structure for storing the definition and status of a submitted or unsubmitted IPUMS data extract request for the purpose of interacting with the IPUMS API.

ipums_extract is a superclass encompassing all of the collection-specific extract classes. Currently supported collections are:

IPUMS USA

IPUMS CPS

IPUMS NHGIS

All objects with class ipums_extract will also have a collection-specific subclass (e.g. usa_extract, cps_extract) to accommodate collection-specific differences in extract options and content. All of these classes share the similarities described below.

Learn more about the IPUMS API in vignette("ipums-api").

Properties

Objects of class ipums_extract have:

A class attribute of the form c("{collection}_extract", "ipums_extract"). For instance, c("cps_extract", "ipums_extract")

A base type of "list"

A names attribute that is a character vector the same length as the underlying list

Creating an extract

Create an ipums_extract object from scratch with the appropriate define_extract function. These functions take the form define_extract_{collection}

Use get_extract_info() to get the latest status of a submitted extract request

Use get_recent_extracts_info() to obtain the extract definitions of previously-submitted extracts

Submitting an extract

Use submit_extract() to submit an extract request for processing through the IPUMS API

wait_for_extract() will periodically check the status of a submitted extract request until it is ready to download

Use is_extract_ready() to manually check whether a submitted extract request is ready to download

Downloading an extract

Download the data contained in a completed extract with download_extract()

Revising an extract definition

Modify values in an existing extract definition with add_to_extract() and remove_from_extract()

Combine definitions from multiple extracts into a single extract with combine_extracts()

Saving an extract

Save an extract to a JSON-formatted file with save_extract_as_json()

Create an ipums_extract object from a saved JSON-formatted definition with define_extract_from_json()