List files contained within a zipped IPUMS extract

Identify the files that can be read from an IPUMS extract.

Usage

ipums_list_files(
  file,
  file_select = NULL,
  types = NULL,
  data_layer = deprecated(),
  shape_layer = deprecated(),
  raster_layer = deprecated()
)

Arguments

file

Path to a .zip archive containing the IPUMS extract to be examined.

file_select

If the path in file contains multiple files, a tidyselect selection identifying the files to be included in the output. Only files that match the provided expression will be included.

While less useful, this can also be provided as a string specifying an exact file name or an integer to match files by index position.

types

One or more of "data" or "shape" indicating the type of files to include in the output. "data" refers to tabular data sources, while "shape" refers to spatial data sources.

The use of "raster" has been deprecated and will be removed in a future release.

data_layer, shape_layer, raster_layer

Please use file_select instead.

Value

A tibble containing the types and names of the available files.

See also

Examples

nhgis_file <- ipums_example("nhgis0712_csv.zip")

# 2 available files in this extract
ipums_list_files(nhgis_file)
#> # A tibble: 2 × 2
#>   type  file                                       
#>   <chr> <chr>                                      
#> 1 data  nhgis0712_csv/nhgis0712_ds135_1990_pmsa.csv
#> 2 data  nhgis0712_csv/nhgis0712_ds136_1990_pmsa.csv

# Look for files that match a particular pattern:
ipums_list_files(nhgis_file, file_select = matches("ds136"))
#> # A tibble: 1 × 2
#>   type  file                                       
#>   <chr> <chr>                                      
#> 1 data  nhgis0712_csv/nhgis0712_ds136_1990_pmsa.csv