Get contextual information about variables in an IPUMS data source

Summarize the variable metadata for the variables found in an ipums_ddi object or data frame. Provides descriptions of variable content (var_label and var_desc) as well as labels for particular values in each variable (val_labels).

ipums_var_info() produces a tibble summary of multiple variables at once.

ipums_var_label(), ipums_var_desc(), and ipums_val_labels() provide specific metadata for a single variable.

ipums_var_info(object, vars = NULL)

ipums_var_label(object, var = NULL)

ipums_var_desc(object, var = NULL)

ipums_val_labels(object, var = NULL)

Arguments

object

An ipums_ddi object, a data frame containing variable metadata (as produced by most ipumsr data-reading functions), or a haven::labelled() vector from a single column in such a data frame.

vars, var

A tidyselect selection identifying the variable(s) to include in the output. Only ipums_var_info() allows for the selection of multiple variables.

Value

For ipums_var_info(), a tibble containing variable information.

Otherwise, a length-1 character vector with the requested variable information.

Details

For ipums_var_info(), if the provided object is a haven::labelled() vector (i.e. a single column from a data frame), the summary output will include the variable label, variable description, and value labels, if applicable.

If it is a data frame, the same information will be provided for all variables present in the data or to those indicated in vars.

If it is an ipums_ddi object, the summary will also include information used when reading the data from disk, including start/end positions for columns in the fixed-width file, implied decimals, and variable types.

Providing an ipums_ddi object is the most robust way to access variable metadata, as many data processing operations will remove these attributes from data frame-like objects.

See also

Examples

ddi <- read_ipums_ddi(ipums_example("cps_00006.xml"))

# Info for all variables in a data source
ipums_var_info(ddi)
#> # A tibble: 8 × 10
#>   var_name var_la…¹ var_d…² val_la…³ code_…⁴ start   end imp_d…⁵ var_t…⁶ recty…⁷
#>   <chr>    <chr>    <chr>   <list>   <chr>   <dbl> <dbl>   <dbl> <chr>   <lgl>  
#> 1 YEAR     Survey … "YEAR … <tibble> "\nYEA…     1     4       0 numeric NA     
#> 2 SERIAL   Househo… "SERIA… <tibble> "\nSER…     5     9       0 numeric NA     
#> 3 HWTSUPP  Househo… "HWTSU… <tibble> "\nHWT…    10    19       4 numeric NA     
#> 4 STATEFIP State (… "STATE… <tibble>  NA        20    21       0 integer NA     
#> 5 MONTH    Month    "MONTH… <tibble>  NA        22    23       0 integer NA     
#> 6 PERNUM   Person … "PERNU… <tibble> "\nPER…    24    25       0 numeric NA     
#> 7 WTSUPP   Supplem… "WTSUP… <tibble> "\nWTS…    26    35       4 numeric NA     
#> 8 INCTOT   Total p… "INCTO… <tibble> "\n999…    36    43       0 numeric NA     
#> # … with abbreviated variable names ¹​var_label, ²​var_desc, ³​val_labels,
#> #   ⁴​code_instr, ⁵​imp_decim, ⁶​var_type, ⁷​rectypes

# Metadata for individual variables
ipums_var_desc(ddi, MONTH)
#> [1] "MONTH indicates the calendar month of the CPS interview."

ipums_var_label(ddi, MONTH)
#> [1] "Month"

ipums_val_labels(ddi, MONTH)
#> # A tibble: 12 × 2
#>      val lbl      
#>    <dbl> <chr>    
#>  1     1 January  
#>  2     2 February 
#>  3     3 March    
#>  4     4 April    
#>  5     5 May      
#>  6     6 June     
#>  7     7 July     
#>  8     8 August   
#>  9     9 September
#> 10    10 October  
#> 11    11 November 
#> 12    12 December