Read metadata about an IPUMS extract from a DDI codebook (.xml) file

Reads the metadata about an IPUMS extract from a DDI codebook into an ipums_ddi object.

These metadata contains parsing instructions for the associated fixed-width data file, contextual labels for variables and values in the data, and general extract information.

See Downloading IPUMS files below for information about downloading IPUMS DDI codebook files.

read_ipums_ddi(
  ddi_file,
  file_select = NULL,
  lower_vars = FALSE,
  data_layer = deprecated()
)

Arguments

ddi_file

Path to a DDI .xml file downloaded from IPUMS or a .zip archive or directory containing the .xml file.

See Downloading IPUMS files below.

file_select

If ddi_file is a .zip archive or directory that contains multiple .xml files, an expression identifying the file to read. Accepts a character string specifying the file name, a tidyselect selection, or an index position. Ignored if ddi_file is the path to a single .xml file.

lower_vars

Logical indicating whether to convert variable names to lowercase. Defaults to FALSE for consistency with IPUMS conventions.

data_layer

Please use file_select instead.

Value

An ipums_ddi object with metadata information.

Downloading IPUMS files

The DDI codebook (.xml) file provided with IPUMS microdata extracts can be downloaded through the IPUMS extract interface or (for some collections) within R using the IPUMS API.

If using the IPUMS extract interface:

Download the DDI codebook by right clicking on the DDI link in the Codebook column of the extract interface and selecting Save as... (on Safari, you may have to select Download Linked File As...). Be sure that the codebook is downloaded in .xml format.

If using the IPUMS API:

For supported collections, use download_extract() to download a completed extract via the IPUMS API. This automatically downloads both the DDI codebook and the data file from the extract and returns the path to the codebook file.

See also

Examples

# Example codebook file
ddi_file <- ipums_example("cps_00006.xml")

# Load data into an `ipums_ddi` obj
ddi <- read_ipums_ddi(ddi_file)

# Use the object to load its associated data
cps <- read_ipums_micro(ddi)
#> Use of data from IPUMS-CPS is subject to conditions including that users should cite the data appropriately. Use command `ipums_conditions()` for more details.

head(cps)
#> # A tibble: 6 × 8
#>    YEAR SERIAL HWTSUPP STATEFIP       MONTH     PERNUM WTSUPP INCTOT            
#>   <dbl>  <dbl>   <dbl> <int+lbl>      <int+lbl>  <dbl>  <dbl> <dbl+lbl>         
#> 1  1962     80   1476. 55 [Wisconsin] 3 [March]      1  1476.     4883          
#> 2  1962     80   1476. 55 [Wisconsin] 3 [March]      2  1471.     5800          
#> 3  1962     80   1476. 55 [Wisconsin] 3 [March]      3  1579. 99999998 [Missing…
#> 4  1962     82   1598. 27 [Minnesota] 3 [March]      1  1598.    14015          
#> 5  1962     83   1707. 27 [Minnesota] 3 [March]      1  1707.    16552          
#> 6  1962     84   1790. 27 [Minnesota] 3 [March]      1  1790.     6375          

# Or get metadata information directly
ipums_var_info(ddi)
#> # A tibble: 8 × 10
#>   var_name var_la…¹ var_d…² val_la…³ code_…⁴ start   end imp_d…⁵ var_t…⁶ recty…⁷
#>   <chr>    <chr>    <chr>   <list>   <chr>   <dbl> <dbl>   <dbl> <chr>   <lgl>  
#> 1 YEAR     Survey … "YEAR … <tibble> "\nYEA…     1     4       0 numeric NA     
#> 2 SERIAL   Househo… "SERIA… <tibble> "\nSER…     5     9       0 numeric NA     
#> 3 HWTSUPP  Househo… "HWTSU… <tibble> "\nHWT…    10    19       4 numeric NA     
#> 4 STATEFIP State (… "STATE… <tibble>  NA        20    21       0 integer NA     
#> 5 MONTH    Month    "MONTH… <tibble>  NA        22    23       0 integer NA     
#> 6 PERNUM   Person … "PERNU… <tibble> "\nPER…    24    25       0 numeric NA     
#> 7 WTSUPP   Supplem… "WTSUP… <tibble> "\nWTS…    26    35       4 numeric NA     
#> 8 INCTOT   Total p… "INCTO… <tibble> "\n999…    36    43       0 numeric NA     
#> # … with abbreviated variable names ¹​var_label, ²​var_desc, ³​val_labels,
#> #   ⁴​code_instr, ⁵​imp_decim, ⁶​var_type, ⁷​rectypes

ipums_file_info(ddi)[1:2]
#> $ipums_project
#> [1] "IPUMS-CPS"
#> 
#> $extract_date
#> [1] "2017-08-02"
#> 

# If variable metadata have been lost from a data source, reattach from
# its corresponding `ipums_ddi` object:
cps <- zap_ipums_attributes(cps)

ipums_var_label(cps$STATEFIP)
#> [1] NA

cps <- set_ipums_var_attributes(cps, ddi$var_info)

ipums_var_label(cps$STATEFIP)
#> [1] "State (FIPS code)"