Read tabular data from an NHGIS extract

Read a csv or fixed-width (.dat) file downloaded from the NHGIS extract system.

To read spatial data from an NHGIS extract, use read_ipums_sf().

read_nhgis(
  data_file,
  file_select = NULL,
  var_attrs = c("val_labels", "var_label", "var_desc"),
  remove_extra_header = TRUE,
  do_file = NULL,
  file_type = NULL,
  na = NULL,
  col_names = TRUE,
  locale = NULL,
  verbose = TRUE,
  progress = NULL,
  show_col_types = NULL,
  ...,
  data_layer = deprecated()
)

Arguments

data_file

Path to a data file, a .zip archive from an NHGIS extract, or a directory containing the data file.

file_select

If data_file is a .zip archive or directory that contains multiple files, an expression identifying the file to load. Accepts a character vector specifying the file name, a tidyselect selection, or an index position. This must uniquely identify a file.

var_attrs

Variable attributes to add from the codebook (.txt) file included in the extract. Defaults to all available attributes.

If the codebook is not found, data will be loaded with no variable attributes.

See set_ipums_var_attributes() for more details.

remove_extra_header

If TRUE, remove the additional descriptive header row that is included in some NHGIS .csv files. Otherwise, the additional header will appear in the first row of the output data frame.

The header contains similar information to that included in the "label" attribute of each data column (if var_attrs includes "var_label"), and is therefore not usually needed.

do_file

For fixed-width files, path to the .do file associated with the provided data_file. The .do file contains the specifications that indicate how to parse the data file.

By default, looks in the same directory as data_file for a .do file with the same name. If FALSE or if the .do file cannot be found, the .dat file will be parsed by the values provided to col_positions in read_fwf(). See Details section.

file_type

One of "csv" (for csv files) or "dat" (for fixed-width files) indicating the type of file to search for in the path provided to data_file. If NULL, determines the file type automatically based on the files found in data_file. Only needed if data_file contains both .csv and .dat files.

na

Character vector of strings to interpret as missing values. If NULL, defaults to c("", "NA") for csv files and c(".", "", "NA") for fixed-width files. See read_csv().

col_names

If reading a .csv file, either TRUE, FALSE or a character vector of column names.

If TRUE, use the default column names included in NHGIS .csv files. If FALSE or a character vector, replace the default column names with the provided names or default placeholders. Note that any columns that have altered names will not be attached to the requested var_attrs.

Note that unlike readr::read_csv(), the first row of the input (which contains NHGIS default headers) will always be removed from the data.

locale

Controls defaults that vary from place to place. If NULL, uses a locale designed to provide appropriate defaults for NHGIS files; altering the locale may cause problems with data parsing.

If needed, you can use readr::locale() to specify a different locale to control things like default time zone, decimal mark, big mark, and day/month names. If you do so, specify encoding = "latin1" to ensure files are encoded properly.

verbose

Logical controlling whether to display output when loading data. If TRUE, displays IPUMS conditions, a progress bar, and guessed column types, unless progress or show_col_types is specified.

progress

Logical indicating whether to display a progress bar when loading data. By default, uses the value of verbose, unless the readr.show_progress option is set to FALSE.

show_col_types

Logical indicating whether to display guessed column types. By default, uses the value of verbose, unless the readr.show_col_types option is set to FALSE.

...

Additional arguments passed to read_csv() or read_fwf().

data_layer

Please use file_select instead.

Value

A tibble containing the data found in data_file

Details

The .do file that is included when downloading an NHGIS fixed-width extract contains important metadata about the data file's column positions and implicit decimals. read_nhgis() uses this information to parse and recode the fixed-width data appropriately. Therefore, while possible, it is not recommended to read fixed-width files without a corresponding .do file.

If you no longer have access to the .do file, consider resubmitting the extract that produced the data.

See also

Examples

# Example files
csv_file <- ipums_example("nhgis0972_csv.zip")
fw_file <- ipums_example("nhgis0730_fixed.zip")

# Provide the .zip archive directly to load the data inside:
read_nhgis(csv_file)
#> Use of data from NHGIS is subject to conditions including that users should cite the data appropriately. Use command `ipums_conditions()` for more details.
#> Rows: 71 Columns: 25
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (9): GISJOIN, STUSAB, CMSA, PMSA, PMSAA, AREALAND, AREAWAT, ANPSADPI, F...
#> dbl (13): YEAR, MSA_CMSAA, INTPTLAT, INTPTLNG, PSADC, D6Z001, D6Z002, D6Z003...
#> lgl  (3): DIVISIONA, REGIONA, STATEA
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 71 × 25
#>    GISJOIN  YEAR STUSAB CMSA  DIVIS…¹ MSA_C…² PMSA  PMSAA REGIONA STATEA AREAL…³
#>    <chr>   <dbl> <chr>  <chr> <lgl>     <dbl> <chr> <chr> <lgl>   <lgl>  <chr>  
#>  1 G0080    1990 OH     28    NA         1692 Akro… 0080  NA      NA     000234…
#>  2 G0360    1990 CA     49    NA         4472 Anah… 0360  NA      NA     000204…
#>  3 G0440    1990 MI     35    NA         2162 Ann … 0440  NA      NA     000183…
#>  4 G0620    1990 IL     14    NA         1602 Auro… 0620  NA      NA     000217…
#>  5 G0845    1990 PA     78    NA         6282 Beav… 0845  NA      NA     000112…
#>  6 G0875    1990 NJ     70    NA         5602 Berg… 0875  NA      NA     000108…
#>  7 G1120    1990 MA     07    NA         1122 Bost… 1120  NA      NA     000455…
#>  8 G1125    1990 CO     34    NA         2082 Boul… 1125  NA      NA     000192…
#>  9 G1145    1990 TX     42    NA         3362 Braz… 1145  NA      NA     000359…
#> 10 G1160    1990 CT     70    NA         5602 Brid… 1160  NA      NA     000067…
#> # … with 61 more rows, 14 more variables: AREAWAT <chr>, ANPSADPI <chr>,
#> #   FUNCSTAT <chr>, INTPTLAT <dbl>, INTPTLNG <dbl>, PSADC <dbl>, D6Z001 <dbl>,
#> #   D6Z002 <dbl>, D6Z003 <dbl>, D6Z004 <dbl>, D6Z005 <dbl>, D6Z006 <dbl>,
#> #   D6Z007 <dbl>, D6Z008 <dbl>, and abbreviated variable names ¹​DIVISIONA,
#> #   ²​MSA_CMSAA, ³​AREALAND

# For extracts that contain multiple files, use `file_select` to specify
# a single file to load.

# This accepts a tidyselect expression
read_nhgis(fw_file, file_select = matches("ds239"), verbose = FALSE)
#> # A tibble: 1 × 114
#>   YEAR    STUSAB NATION NATIONA AIHHTLI MEMI  PCI   GEOID NAME_E AJWBE…¹ AJWBE…²
#>   <chr>   <chr>  <chr>  <chr>   <chr>   <chr> <chr> <chr> <chr>    <dbl>   <dbl>
#> 1 2014-2… US     Unite… 1       NA      NA    NA    0100… Unite…  3.23e8  1.59e8
#> # … with 103 more variables: AJWBE003 <dbl>, AJWBE004 <dbl>, AJWBE005 <dbl>,
#> #   AJWBE006 <dbl>, AJWBE007 <dbl>, AJWBE008 <dbl>, AJWBE009 <dbl>,
#> #   AJWBE010 <dbl>, AJWBE011 <dbl>, AJWBE012 <dbl>, AJWBE013 <dbl>,
#> #   AJWBE014 <dbl>, AJWBE015 <dbl>, AJWBE016 <dbl>, AJWBE017 <dbl>,
#> #   AJWBE018 <dbl>, AJWBE019 <dbl>, AJWBE020 <dbl>, AJWBE021 <dbl>,
#> #   AJWBE022 <dbl>, AJWBE023 <dbl>, AJWBE024 <dbl>, AJWBE025 <dbl>,
#> #   AJWBE026 <dbl>, AJWBE027 <dbl>, AJWBE028 <dbl>, AJWBE029 <dbl>, …

# Or an index position
read_nhgis(fw_file, file_select = 2, verbose = FALSE)
#> # A tibble: 84 × 28
#>    GISJOIN STATE STATEFP STATENH A00AA…¹ A00AA…² A00AA…³ A00AA…⁴ A00AA…⁵ A00AA…⁶
#>    <chr>   <chr> <chr>   <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 G010    Alab… 01      010          NA      NA      NA  127901  309527  590756
#>  2 G020    Alas… 02      020          NA      NA      NA      NA      NA      NA
#>  3 G025    Alas… NA      025          NA      NA      NA      NA      NA      NA
#>  4 G040    Ariz… 04      040          NA      NA      NA      NA      NA      NA
#>  5 G045    Ariz… NA      045          NA      NA      NA      NA      NA      NA
#>  6 G050    Arka… 05      050          NA      NA      NA      NA      NA   97574
#>  7 G055    Arka… NA      055          NA      NA      NA   14273   30388      NA
#>  8 G060    Cali… 06      060          NA      NA      NA      NA      NA      NA
#>  9 G080    Colo… 08      080          NA      NA      NA      NA      NA      NA
#> 10 G085    Colo… NA      085          NA      NA      NA      NA      NA      NA
#> # … with 74 more rows, 18 more variables: A00AA1850 <dbl>, A00AA1860 <dbl>,
#> #   A00AA1870 <dbl>, A00AA1880 <dbl>, A00AA1890 <dbl>, A00AA1900 <dbl>,
#> #   A00AA1910 <dbl>, A00AA1920 <dbl>, A00AA1930 <dbl>, A00AA1940 <dbl>,
#> #   A00AA1950 <dbl>, A00AA1960 <dbl>, A00AA1970 <dbl>, A00AA1980 <dbl>,
#> #   A00AA1990 <dbl>, A00AA2000 <dbl>, A00AA2010 <dbl>, A00AA2020 <dbl>, and
#> #   abbreviated variable names ¹​A00AA1790, ²​A00AA1800, ³​A00AA1810, ⁴​A00AA1820,
#> #   ⁵​A00AA1830, ⁶​A00AA1840

# NHGIS fixed-width files use the information contained in the extract's
# .do file to parse the data correctly. If it does not exist, the parsing
# will be incorrect. If you have moved the .do file, be sure to provide
# its new file path to the `do_file` argument.
bad_parse <- read_nhgis(
  fw_file,
  file_select = 2,
  do_file = FALSE,
  verbose = FALSE
)
#> Warning: Data loaded from NHGIS fixed-width files may not be consistent with the information included in the data codebook when parsing column positions manually.
#> ℹ Please consult the .txt and .do files associated with this extract to ensure data is recoded correctly.

bad_parse$X1[1:10]
#>  [1] "G010Alabama"            "G020Alaska"             "G025Alaska Territory"  
#>  [4] "G040Arizona"            "G045Arizona Territory"  "G050Arkansas"          
#>  [7] "G055Arkansas Territory" "G060California"         "G080Colorado"          
#> [10] "G085Colorado Territory"

# `readr::read_csv()` arguments are accepted, though most defaults should
# be appropriate.
read_nhgis(csv_file, col_types = readr::cols(.default = "c"))
#> Use of data from NHGIS is subject to conditions including that users should cite the data appropriately. Use command `ipums_conditions()` for more details.
#> Rows: 71 Columns: 25
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (25): GISJOIN, YEAR, STUSAB, CMSA, DIVISIONA, MSA_CMSAA, PMSA, PMSAA, RE...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 71 × 25
#>    GISJOIN YEAR  STUSAB CMSA  DIVIS…¹ MSA_C…² PMSA  PMSAA REGIONA STATEA AREAL…³
#>    <chr>   <chr> <chr>  <chr> <chr>   <chr>   <chr> <chr> <chr>   <chr>  <chr>  
#>  1 G0080   1990  OH     28    NA      1692    Akro… 0080  NA      NA     000234…
#>  2 G0360   1990  CA     49    NA      4472    Anah… 0360  NA      NA     000204…
#>  3 G0440   1990  MI     35    NA      2162    Ann … 0440  NA      NA     000183…
#>  4 G0620   1990  IL     14    NA      1602    Auro… 0620  NA      NA     000217…
#>  5 G0845   1990  PA     78    NA      6282    Beav… 0845  NA      NA     000112…
#>  6 G0875   1990  NJ     70    NA      5602    Berg… 0875  NA      NA     000108…
#>  7 G1120   1990  MA     07    NA      1122    Bost… 1120  NA      NA     000455…
#>  8 G1125   1990  CO     34    NA      2082    Boul… 1125  NA      NA     000192…
#>  9 G1145   1990  TX     42    NA      3362    Braz… 1145  NA      NA     000359…
#> 10 G1160   1990  CT     70    NA      5602    Brid… 1160  NA      NA     000067…
#> # … with 61 more rows, 14 more variables: AREAWAT <chr>, ANPSADPI <chr>,
#> #   FUNCSTAT <chr>, INTPTLAT <chr>, INTPTLNG <chr>, PSADC <chr>, D6Z001 <chr>,
#> #   D6Z002 <chr>, D6Z003 <chr>, D6Z004 <chr>, D6Z005 <chr>, D6Z006 <chr>,
#> #   D6Z007 <chr>, D6Z008 <chr>, and abbreviated variable names ¹​DIVISIONA,
#> #   ²​MSA_CMSAA, ³​AREALAND