Remove values from an existing NHGIS extract definition

Remove existing values from an IPUMS NHGIS extract definition. All fields are optional, and if omitted, will be unchanged.

In general, removing from an extract follows the same syntax conventions as used in define_extract_nhgis(). See Details section for more information on how values passed to dataset and time series table subfields are interpreted.

To add new values to an IPUMS NHGIS extract definition, use add_to_extract(). When replacing values, it is best to first add new values using add_to_extract() before removing values with remove_from_extract(). This limits the possibility of producing a temporarily invalid extract specification.

Learn more about the IPUMS API in vignette("ipums-api").

# S3 method for nhgis_extract
remove_from_extract(
  extract,
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  ...
)

Arguments

extract

An ipums_extract object.

datasets

Datasets to remove from the extract definition. All data_tables, geog_levels, years, and breakdown_values associated with the specified datasets will also be removed.

data_tables

Summary tables to remove from the datasets in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets.

time_series_tables

Time series tables to remove from the extract definition. All geog_levels associated with the specified time_series_tables will also be removed.

geog_levels

Geographic levels to remove from the datasets and/or time_series_tables in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets or time_series_tables.

years

Years to remove from the datasets in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets.

breakdown_values

Breakdown values to remove from the datasets in the extract definition.

See Details section for syntax options when an extract definition has multiple datasets..

geographic_extents

Geographic extents to remove from the extract definition.

shapefiles

Shapefiles to remove from the extract definition.

...

Ignored

Value

A modified nhgis_extract object

Details

For extract definitions that include any datasets, corresponding data_tables and geog_levels must be specified. For extract definitions that include any time_series_tables, corresponding geog_levels must be specified.

NHGIS extract definitions may contain multiple datasets or time_series_tables. Each dataset or time series table is associated with several subfields that apply only to that particular dataset or time series table.

The following arguments are removed from the extract definition's datasets:

data_tables

years

breakdown_values

The following arguments are removed from the extract definition's datasets and time_series_tables:

geog_levels

To remove the values of these arguments from all of the datasets and/or time_series_tables in the extract definition, pass the values as a vector.

To remove the values of these arguments to specific datasets and/or time_series_tables in the extract definition, pass the values as a named list. In this case, names should correspond to the datasets and/or time_series_tables and their associated values will only be removed from the indicated dataset or time series table.

Note that subfields are modified after the removal of any datasets and time_series_tables specified in datasets or time_series_tables arguments.

While while it is possible to match elements to datasets and/or time_series_tables by index by providing an unnamed list, we recommend using names to explicitly specify the values that should be removed from each of the request's datasets and time_series_tables.

Any extract fields that are rendered irrelevant after modifying the extract will be automatically removed. (For instance, if all time_series_tables are removed from an extract, tst_layout will also be removed.) Thus, it is not necessary to explicitly remove these values.

Note that it is possible to produce invalid extracts using remove_from_extract() (for instance, an extract that includes a time series table without associated geographic levels). This can occur if you intend to replace the existing values for a required extract field. If the goal is not to remove values but to replace them, it is best to first use add_to_extract(), then remove_from_extract(), as this will avoid the possibility of temporarily producing an invalid extract.

If the supplied extract definition comes from a previously submitted extract request, this function will reset the definition to an unsubmitted state.

See also

Examples

extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2", "NP3"),
  geog_levels = "county"
)

# Remove a table from existing datasets:
remove_from_extract(extract, data_tables = "NP3")
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 1990_STF1
#>   Tables: NP1, NP2
#>   Geog Levels: county

extract2 <- define_extract_nhgis(
  time_series_tables = c("CW3", "CW5"),
  geog_levels = c("state", "county")
)

# Remove an entire time series table
remove_from_extract(
  extract2,
  time_series_tables = "CW3"
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Time Series Table: CW5
#>   Geog Levels: state, county

# Use a named list to remove values from specific datasets or time
# series tables:
remove_from_extract(
  extract2,
  geog_levels = list(CW5 = "state", CW3 = "county")
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Time Series Table: CW3
#>   Geog Levels: state
#> 
#> Time Series Table: CW5
#>   Geog Levels: county

# Replace values by using `add_to_extract()` first to avoid invalid
# extract definitions
revised <- add_to_extract(extract2, geog_levels = list(CW3 = "tract"))
remove_from_extract(revised, geog_levels = list(CW3 = c("county", "state")))
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Time Series Table: CW3
#>   Geog Levels: tract
#> 
#> Time Series Table: CW5
#>   Geog Levels: state, county