Wait for an extract request to finish processing

Wait for an extract request to finish by periodically checking its status via the IPUMS API until it is complete.

Completed extracts will have a value of "completed" in their status field.

Learn more about the IPUMS API in vignette("ipums-api").

wait_for_extract(
  extract,
  initial_delay_seconds = 0,
  max_delay_seconds = 300,
  timeout_seconds = 10800,
  verbose = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

Arguments

extract

One of:

An ipums_extract object

The data collection and extract number formatted as a string of the form "collection:number" or as a vector of the form c("collection", number)

An extract number to be associated with your default IPUMS collection. See set_ipums_default_collection()

Extract numbers do not need to be zero-padded. That is, use 1, not "0001".

For a list of codes used to refer to each collection, see ipums_data_collections().

initial_delay_seconds

Seconds to wait before first status check. The wait time will automatically increase by 10 seconds between each successive check.

max_delay_seconds

Maximum interval to wait between status checks. When the wait interval reaches this value, checks will continue to occur at max_delay_seconds intervals until the extract is complete or timeout_seconds is reached. Defaults to 300 seconds (5 minutes).

timeout_seconds

Maximum total number of seconds to continue waiting for the extract before throwing an error. Defaults to 10,800 seconds (3 hours).

verbose

If TRUE, print status updates to the R console at the beginning of each wait interval and upon extract completion. Defaults to TRUE.

api_key

API key associated with your user account. Defaults to the value of the IPUMS_API_KEY environment variable. See set_ipums_api_key().

Value

An ipums_extract object containing the extract definition and the URLs from which to download extract files.

See also

Examples

my_extract <- define_extract_usa(
  description = "2013-2014 ACS Data",
  samples = c("us2013a", "us2014a"),
  variables = c("SEX", "AGE", "YEAR")
)

if (FALSE) {
submitted_extract <- submit_extract(my_extract)

# Wait for a particular extract request to complete by providing its
# associated `ipums_extract` object:
downloadable_extract <- wait_for_extract(submitted_extract)

# Or by specifying the collection and number for the extract request:
downloadable_extract <- wait_for_extract("usa:1")

# If you have a default collection, you can use the extract number alone:
set_ipums_default_collection("usa")
downloadable_extract <- wait_for_extract(1)

# If you want to subsequently download your completed extract, you can
# use `download_extract()` directly:
download_extract(1, wait = TRUE)
}